scripts/automation/requirements.txt
# Core dependencies
pyyaml>=6.0.1
python-dotenv>=1.0.0
aiohttp>=3.9.0
httpx>=0.27.0
typer>=0.12.0
aiosqlite>=0.19.0

# Testing
pytest>=9.0.0
pytest-asyncio>=1.3.0
pytest-cov>=5.0.0

# Slack integration (optional - using webhook URL instead of SDK)
# slack-sdk>=3.23.0
